Maint:use float for data length

+++ evaluation.py
@@ -6,7 +6,7 @@ pd.options.mode.chained_assignment = None  # default='warn'
 ### Parser
 parser = argparse.ArgumentParser(description='Add some integers.')
 
-parser.add_argument('--length', action='store', type=int, default=5,
+parser.add_argument('--length', action='store', type=float, default=5,
                     help='Length of data to take into account (0,5].')
 
 parser.add_argument('--subjects', action='store', type=int, default=35,
